linkers/: a few long->int32 for lput, wput, to fix clang macos compilation

+++ linkers/5l/asm.c
@@ -261,7 +261,7 @@ cput(int c)
 }
 
 void
-wput(long l)
+wput(int32 l)
 {
 
 	cbp[0] = l>>8;
@@ -273,7 +273,7 @@ wput(long l)
 }
 
 void
-wputl(long l)
+wputl(int32 l)
 {
 
 	cbp[0] = l;
@@ -285,7 +285,7 @@ wputl(long l)
 }
 
 void
-lput(long l)
+lput(int32 l)
 {
 
 	cbp[0] = l>>24;
@@ -299,7 +299,7 @@ lput(long l)
 }
 
 void
-lputl(long l)
+lputl(int32 l)
 {
 
 	cbp[3] = l>>24;
