+++ linkers/8l/asm.c
@@ -30,21 +30,21 @@ entryvalue(void)
 
 /* these need to take long arguments to be compatible with elf.c */
 void
-wputl(long w)
+wputl(int32 w)
 {
 	cput(w);
 	cput(w>>8);
 }
 
 void
-wput(long w)
+wput(int32 w)
 {
 	cput(w>>8);
 	cput(w);
 }
 
 void
-lput(long l)
+lput(int32 l)
 {
 	cput(l>>24);
 	cput(l>>16);
@@ -53,7 +53,7 @@ lput(long l)
 }
 
 void
-lputl(long l)
+lputl(int32 l)
 {
 	cput(l);
 	cput(l>>8);
